Moving BuildCustomRecipeItemDescription into CraftMagicItemsBlueprintPatcher as it is the only code referencing the method
